refactor: service 계층 제거
